Tidy everything up
