+++ Src/java/cql-to-elm/src/main/java/org/cqframework/cql/cql2elm/elm/ElmEditor.java
@@ -10,10 +10,10 @@ import org.hl7.elm.r1.Library;
 
 public class ElmEditor {
 
-    private final List<ElmEdit> edits;
+    private final List<IElmEdit> edits;
     private final FunctionalElmVisitor<Trackable, Void> visitor;
 
-    public ElmEditor(List<ElmEdit> edits) {
+    public ElmEditor(List<IElmEdit> edits) {
         this.edits = Objects.requireNonNull(edits);
         this.visitor = Visitors.from((elm, context) -> elm, this::aggregateResults);
     }
@@ -21,23 +21,18 @@ public class ElmEditor {
     public void edit(Library library) {
         this.visitor.visitLibrary(library, null);
 
-        // This is needed because aggregateResults is not called for the library itself.
+        // This is needed because aggregateResults is not called on the library itself.
         this.applyEdits(library);
     }
 
     protected Trackable aggregateResults(Trackable aggregate, Trackable nextResult) {
-        if (nextResult instanceof Element) {
-            for (ElmEdit edit : edits) {
-                edit.edit((Element) nextResult);
-            }
-        }
-
+        applyEdits(nextResult);
         return aggregate;
     }
 
     protected void applyEdits(Trackable trackable) {
         if (trackable instanceof Element) {
-            for (ElmEdit edit : edits) {
+            for (IElmEdit edit : edits) {
                 edit.edit((Element) trackable);
             }
         }
