+++ Src/java/cql-to-elm/src/main/java/org/cqframework/cql/cql2elm/elm/IElmEdit.java
@@ -1,0 +1,7 @@
+package org.cqframework.cql.cql2elm.elm;
+
+import org.hl7.elm.r1.Element;
+
+public interface IElmEdit {
+    void edit(Element element);
+}
